+++ block-sys/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Fixed
+* **BREAKING**: `Class` is now `!UnwindSafe`.
 
 ## 0.0.1 - 2021-11-22