+++ objc-sys/CHANGELOG.md
@@ -19,6 +19,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   present on other platforms anyhow, so users will just get an error at
   link-time.
 
+### Fixed
+* **BREAKING**: Opaque types are now also `!UnwindSafe`.
+
 
 ## 0.1.0 - 2021-11-22